+++ lib/src/visitor.dart
@@ -77,6 +77,8 @@ abstract class Visitor<C, R> {
 
   R visitTemplateNode(TemplateNode node, C context);
 
+  R visitTryCatch(TryCatch node, C context);
+
   R visitWith(With node, C context);
 
   R visitSlice(Slice node, C context) {
@@ -263,6 +265,11 @@ class ThrowingVisitor<C, R> implements Visitor<C, R> {
 
   @override
   R visitTemplateNode(TemplateNode node, C context) {
+    throw UnimplementedError();
+  }
+
+  @override
+  R visitTryCatch(TryCatch node, C context) {
     throw UnimplementedError();
   }